pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "qr-sampler"
dynamic = ["version"]
description = "Plug any randomness source into LLM token sampling via vLLM"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "qr-sampler contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
readme = "README.md"

[project.urls]
Homepage = "https://github.com/qr-sampler/qr-sampler"
Repository = "https://github.com/qr-sampler/qr-sampler"
Issues = "https://github.com/qr-sampler/qr-sampler/issues"
Changelog = "https://github.com/qr-sampler/qr-sampler/blob/main/CHANGELOG.md"

dependencies = [
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
grpc = [
    "grpcio>=1.60.0",
    "protobuf>=4.21.0",
]
dev = [
    "qr-sampler[grpc]",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "scipy>=1.10.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "pre-commit>=3.0",
    "bandit>=1.7.0",
]

[project.entry-points."vllm.logits_processors"]
qr_sampler = "qr_sampler.processor:QRSamplerLogitsProcessor"

[project.entry-points."qr_sampler.entropy_sources"]
system = "qr_sampler.entropy.system:SystemEntropySource"
quantum_grpc = "qr_sampler.entropy.quantum:QuantumGrpcSource"
timing_noise = "qr_sampler.entropy.timing:TimingNoiseSource"
mock_uniform = "qr_sampler.entropy.mock:MockUniformSource"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
fallback_version = "0.1.0.dev0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",       # pyflakes + pycodestyle
    "I",                  # isort
    "N",                  # pep8-naming
    "UP",                 # pyupgrade
    "B",                  # flake8-bugbear
    "SIM",                # flake8-simplify
    "TCH",                # flake8-type-checking
    "RUF",                # ruff-specific rules
]

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true
packages = ["qr_sampler"]
mypy_path = "src"

[tool.coverage.run]
source = ["src/qr_sampler"]
omit = [
    "src/qr_sampler/proto/*",
]

[tool.coverage.report]
fail_under = 90
show_missing = true
